requirements.txt
cryptography==38.0.1
dj-database-url==0.5.0
Django>=2.2.20
django-markdown-deux==1.0.5
django-tinymce==2.8.0
djangorestframework==3.11.2
gunicorn==19.9.0
markdown2==2.3.7
mysql-connector-python==8.0.15
protobuf==3.6.1
psycopg2==2.9.2
pycparser==2.19
PyMySQL==0.9.3
pytz==2018.9
six==1.12.0
whitenoise==5.3.0
